+++ NEWS.md
@@ -15,6 +15,7 @@ Version 0.18b1 - dev
   `add_ordinate_dim()`, `add_ordinate_x_dim()`, `add_ordinate_y_dim()`
 - NEW: function `ezdxf.tools.text.is_upside_down_text_angle()` in WCS
 - NEW: function `ezdxf.tools.text.upright_text_angle()` in WCS
+- NEW: extended query functionality for the `EntityQuery` class
 - NEW: helper class `ezdxf.math.ConstructionPolyline` to measure, interpolate and 
   divide polylines and anything that can be approximated or flattened into 
   vertices